appointment-service/requirements.txt
flask==3.0.0
pymongo==4.6.0
PyJWT==2.8.0
flask-cors==4.0.0
